register read ok

+++ chipTest/app/bm1794/reg-scan.h
@@ -9,7 +9,7 @@ struct reg_scan_item_t
 {
 	uint8_t chain_id;
 	uint8_t chip_addr;
-	uint8_t reg_addr;
+	uint32_t reg_addr;
 	uint32_t reg_data;
 	uint8_t age;
 };
